+++ src/project/components/app/app.tsx
@@ -1,9 +1,9 @@
 import { MainPage } from '../../pages/main-page/main-page';
-import { RentCount } from '../../const/const';
+import { RentQuantity } from '../../const/const';
 
 function App() {
   return (
-    <MainPage props={RentCount.count} />
+    <MainPage props={RentQuantity.quantity} />
   );
 }
