+++ src/project/types/index.ts
@@ -1,0 +1,2 @@
+export from './Comments.type';
+export from './Offers.type'
