+++ src/project/utils/utils.ts
@@ -1,0 +1,12 @@
+const PLACE_COUNT_DEFAULT = 6;
+
+function getIndexValue<T extends number>(index: T){
+  return {
+    id: index
+  };
+}
+
+const placeCountArr = Array.from({length: PLACE_COUNT_DEFAULT}, (_, index,) => getIndexValue(index));
+
+
+export { placeCountArr};
